requirements.txt
transformers==4.36.0
torch==2.1.0
datasets==2.16.0
scikit-learn==1.3.0
shap==0.44.0
streamlit==1.29.0
plotly==5.18.0
pandas==2.1.0
numpy==1.26.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
requests==2.31.0
langchain==0.2.16
langchain-core==0.2.38
langchain-community==0.2.16
langchain-groq
langchain-google-genai
tavily-python
